refactor: rename withoutProjectParserOptions file as well
