+++ packages/typescript-estree/src/withoutProjectParserOptions.ts
